refactor(services): split text-diff into focused modules
Break the 956-line text-diff.ts monolith into a directory of focused
modules. The public surface is preserved via index.ts re-exports so
callers continue to import from '@/lib/services/text-diff' unchanged.

New layout:
- types.ts          all type / interface definitions
- options.ts        defaultDiffOptions + normalizeForComparison
- samples.ts        SAMPLE_LEFT_TEXT / SAMPLE_RIGHT_TEXT
- line-diff.ts      computeDiff + getUnifiedDiff + areTextsIdentical
                    plus line-level LCS helpers and buildDiffResult
- char-diff.ts      computeInlineDiff + character LCS helpers
- enhanced-diff.ts  computeEnhancedDiff + formatHunkHeader + hunk grouping
- styles.ts         Tailwind class helpers for the diff viewer
- index.ts          re-export of the public surface

Rationale: line / char / enhanced diffs have distinct concerns and
distinct algorithmic boundaries. Splitting them gives each file a
single responsibility (60-320 LOC) and makes future per-level tuning
(e.g. swapping in Myers diff for lines without touching char diff)
straightforward.

No API changes. diff-viewer.tsx (the sole consumer) is unaffected.
